Task4

+++ Tyuiu.ParazyanAI.Sprint0.Task4.V0.Lib/DataService.cs
@@ -13,16 +13,20 @@ namespace Tyuiu.ParazyanAI.Sprint0.Task4.V0.Lib
             return a + b; 
             }
 
+
         public static int Subtraction(int a, int b)
             { 
             return a - b; 
             }
 
+
         public static int Multiplication(int a, int b)
         { 
             return a * b; 
         }
-        public static int Division(int a, int b)
+
+
+        public static double Division(int a, int b)
         {
             return a / b;
         }
